Правим конфигурирования дизайнов. Частично сделано: 1) механизм выбора дизайна из списка 2) реализация в нескольких индикаторах Надо еще: 3) Реализовать этот механизм в остальных индикаторах. 4) Цветные индикаторы

+++ app/src/main/java/ru/cpc/smartflatview/Door.java
@@ -97,7 +97,7 @@ public class Door extends Indicator
 	}
 
 	@Override
-	public boolean SetValue(int iValue) 
+	public boolean SetValue(float iValue)
 	{
 		//if(iValue > 0)
 		{
