+++ app/src/main/java/ru/cpc/smartflatview/WarmFloorDevi.java
@@ -56,7 +56,7 @@ public class WarmFloorDevi extends BaseRegulator
                 SwitchOnOff(0, 0);
             }
         });
-        final ScrollingDialog.SFSeeker pSeeker = (ScrollingDialog.SFSeeker)ScrollingDialog.AddSeekBar(m_sVariableValue, context.getString(R.string.sdTemperature), m_iValue, (int) m_fValueMin, (int) m_fValueMax, "%d °C", m_iReaction != 0
+        final ScrollingDialog.SFSeeker pSeeker = (ScrollingDialog.SFSeeker)ScrollingDialog.AddSeekBar(m_sVariableValue, context.getString(R.string.sdTemperature),(int) m_iValue, (int) m_fValueMin, (int) m_fValueMax, "%d °C", m_iReaction != 0
                                                                                                                                                                                                                             ? null : new SeekBar.OnSeekBarChangeListener() {
             @Override
             public void onProgressChanged(SeekBar seekBar, int progress, boolean fromUser) {
